requirements.txt
pandas
python-docx
openpyxl
pyinstaller
pillow 